# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: "TechSonance CMS API"
  version: "1.0.0"
  description: "Generated OpenAPI specification for the TechSonance CMS routes."
servers:
  - 
    url: "/"
...
paths:
...
  /api/vendors:
    get:
      tags:
        - vendors
      summary: "Retrieve resource for /api/vendors"
      description: "Auto-generated documentation for GET /api/vendors."
      security:
        - 
          bearerAuth: []
      parameters:
        - 
          name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
          example: 10
        - 
          name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
            minimum: 0
          example: 0
      responses:
        200:
          $ref: "#/components/responses/SuccessResponse"
        201:
          $ref: "#/components/responses/CreatedResponse"
        400:
          $ref: "#/components/responses/ValidationErrorResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
        403:
          $ref: "#/components/responses/ForbiddenResponse"
        404:
          $ref: "#/components/responses/NotFoundResponse"
        422:
          $ref: "#/components/responses/UnprocessableResponse"
        500:
          $ref: "#/components/responses/InternalErrorResponse"
    post:
      tags:
        - vendors
      summary: "Create resource for /api/vendors"
      description: "Auto-generated documentation for POST /api/vendors."
      security:
        - 
          bearerAuth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericPayload"
            examples:
              sample:
                summary: "POST request payload"
                value:
                  name: "Example"
                  status: "active"
      responses:
        201:
          $ref: "#/components/responses/CreatedResponse"
        400:
          $ref: "#/components/responses/ValidationErrorResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
        403:
          $ref: "#/components/responses/ForbiddenResponse"
        404:
          $ref: "#/components/responses/NotFoundResponse"
        422:
          $ref: "#/components/responses/UnprocessableResponse"
        500:
          $ref: "#/components/responses/InternalErrorResponse"
    put:
      tags:
        - vendors
      summary: "Update resource for /api/vendors"
      description: "Auto-generated documentation for PUT /api/vendors."
      security:
        - 
          bearerAuth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericPayload"
            examples:
              sample:
                summary: "PUT request payload"
                value:
                  name: "Example"
                  status: "active"
      responses:
        200:
          $ref: "#/components/responses/SuccessResponse"
        201:
          $ref: "#/components/responses/CreatedResponse"
        400:
          $ref: "#/components/responses/ValidationErrorResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
        403:
          $ref: "#/components/responses/ForbiddenResponse"
        404:
          $ref: "#/components/responses/NotFoundResponse"
        422:
          $ref: "#/components/responses/UnprocessableResponse"
        500:
          $ref: "#/components/responses/InternalErrorResponse"
    delete:
      tags:
        - vendors
      summary: "Delete resource for /api/vendors"
      description: "Auto-generated documentation for DELETE /api/vendors."
      security:
        - 
          bearerAuth: []
      parameters: []
      responses:
        200:
          $ref: "#/components/responses/SuccessResponse"
        400:
          $ref: "#/components/responses/ValidationErrorResponse"
        401:
          $ref: "#/components/responses/UnauthorizedResponse"
        403:
          $ref: "#/components/responses/ForbiddenResponse"
        404:
          $ref: "#/components/responses/NotFoundResponse"
        422:
          $ref: "#/components/responses/UnprocessableResponse"
        500:
          $ref: "#/components/responses/InternalErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
    GenericPayload:
      type: "object"
      additionalProperties: true
      example:
        name: "Example"
        status: "active"
    ApiEnvelope:
      type: "object"
      properties:
        success:
          type: "boolean"
        data:

        message:
          type: "string"
        errors:
          type: "array"
          items:

          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    PaginationMeta:
      type: "object"
      properties:
        page:
          type: "integer"
        limit:
          type: "integer"
        total:
          type: "integer"
  responses:
    SuccessResponse:
      description: "Successful response"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
          examples:
            sample:
              value:
                success: true
                data:
                  id: 1
                message: "Request completed successfully"
                errors: null
    CreatedResponse:
      description: "Created response"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
          examples:
            sample:
              value:
                success: true
                data:
                  id: 1
                message: "Resource created successfully"
                errors: null
    ValidationErrorResponse:
      description: "Validation error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
          examples:
            sample:
              value:
                success: false
                data: null
                message: "Validation failed"
                errors:
                  - 
                    path:
                      - name
                    message: "Required"
    UnauthorizedResponse:
      description: "Unauthorized"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
    ForbiddenResponse:
      description: "Forbidden"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
    NotFoundResponse:
      description: "Not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
    UnprocessableResponse:
      description: "Unprocessable entity"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
    InternalErrorResponse:
      description: "Internal server error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"